+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 v0.2.0 (in development)
 -----------------------
+- Sleep between mutating API requests in order to keep them at least one second
+  apart, as recommended by GitHub
+- Retry (with exponential backoff) requests that fail with 5xx errors or due to
+  rate limiting
 
 v0.1.0 (2023-10-19)
 -------------------